+++ aquadoggo/src/network/config.rs
@@ -11,51 +11,66 @@ pub const NODE_NAMESPACE: &str = "aquadoggo";
 /// Network config for the node.
 #[derive(Debug, Clone)]
 pub struct NetworkConfiguration {
-    /// QUIC port for node-to-node communication.
+    /// QUIC port for node-node communication and data replication.
     pub quic_port: u16,
 
     /// Discover peers on the local network via mDNS (over IPv4 only, using port 5353).
     pub mdns: bool,
 
-    /// List of known node addresses (IP + port) we want to connect to directly.
+    /// List of known node addresses we want to connect to directly.
     ///
-    /// Make sure that nodes mentioned in this list are directly reachable (for example they need
-    /// to be hosted with a static IP Address). If you need to connect to nodes with changing,
-    /// dynamic IP addresses or even with nodes behind a firewall or NAT, do not use this field but
-    /// use at least one relay.
+    /// Make sure that nodes mentioned in this list are directly reachable (they need to be hosted
+    /// with a static IP Address). If you need to connect to nodes with changing, dynamic IP
+    /// addresses or even with nodes behind a firewall or NAT, do not use this field but use at
+    /// least one relay.
     pub direct_node_addresses: Vec<Multiaddr>,
 
-    /// Set to true if node should also function as a relay. Other nodes can use relays to aid
-    /// discovery and establishing connectivity.
+    /// List of peers which are allowed to connect to your node.
+    ///
+    /// If set then only nodes (identified by their peer id) contained in this list will be able to
+    /// connect to your node (via a relay or directly). When not set any other node can connect to
+    /// yours.
+    ///
+    /// Peer IDs identify nodes by using their hashed public keys. They do _not_ represent authored
+    /// data from clients and are only used to authenticate nodes towards each other during
+    /// networking.
+    ///
+    /// Use this list for example for setups where the identifier of the nodes you want to form a
+    /// network with is known but you still need to use relays as their IP addresses change
+    /// dynamically.
+    pub allow_peer_ids: AllowList<PeerId>,
+
+    /// List of peers which will be blocked from connecting to your node.
+    ///
+    /// If set then any peers (identified by their peer id) contained in this list will be blocked
+    /// from connecting to your node (via a relay or directly). When an empty list is provided then
+    /// there are no restrictions on which nodes can connect to yours.
+    ///
+    /// Block lists and allow lists are exclusive, which means that you should _either_ use a block
+    /// list _or_ an allow list depending on your setup.
+    ///
+    /// Use this list for example if you want to allow _any_ node to connect to yours _except_ of a
+    /// known number of excluded nodes.
+    pub block_peer_ids: Vec<PeerId>,
+
+    /// List of relay addresses.
+    ///
+    /// A relay helps discover other nodes on the internet (also known as "rendesvouz" or
+    /// "bootstrap" server) and helps establishing direct p2p connections when node is behind a
+    /// firewall or NAT (also known as "holepunching").
+    ///
+    /// WARNING: This will potentially expose your IP address on the network. Do only connect to
+    /// trusted relays or make sure your IP address is hidden via a VPN or proxy if you're
+    /// concerned about leaking your IP.
+    pub relay_addresses: Vec<Multiaddr>,
+
+    /// Enable if node should also function as a relay.
+    ///
+    /// Other nodes can use relays to aid discovery and establishing connectivity.
     ///
     /// Relays _need_ to be hosted in a way where they can be reached directly, for example with a
     /// static IP address through an VPS.
     pub relay_mode: bool,
-
-    /// Addresses of a peers which can act as a relay/rendezvous server.
-    ///
-    /// Relays help discover other nodes on the internet (also known as "rendesvouz" or "bootstrap"
-    /// server) and help establishing direct p2p connections when node is behind a firewall or NAT
-    /// (also known as "holepunching").
-    ///
-    /// When a direct connection is not possible the relay will help to redirect the (encrypted)
-    /// traffic as an intermediary between us and other nodes. The node will contact each server
-    /// and register our IP address for other peers.
-    pub relay_addresses: Vec<Multiaddr>,
-
-    /// List of peers which can connect to our node.
-    ///
-    /// If set then only peers (identified by their peer id) contained in this list will be able
-    /// to connect to our node (via a relay or directly). When not set then there are no
-    /// restrictions on which nodes can connect to ours.
-    pub allow_peer_ids: AllowList<PeerId>,
-
-    /// List of peers which can connect to our node.
-    ///
-    /// If set then only peers (identified by their peer id) contained in this list will be able
-    /// to connect to our node (via a relay or directly). When not set then there are no
-    /// restrictions on which nodes can connect to ours.
-    pub block_peer_ids: AllowList<PeerId>,
 
     /// Notify handler buffer size.
     ///
@@ -101,21 +116,21 @@ pub struct NetworkConfiguration {
 impl Default for NetworkConfiguration {
     fn default() -> Self {
         Self {
+            quic_port: 2022,
+            mdns: true,
+            direct_node_addresses: Vec::new(),
+            allow_peer_ids: AllowList::<PeerId>::Wildcard,
+            block_peer_ids: Vec::new(),
+            relay_addresses: Vec::new(),
+            relay_mode: false,
+            notify_handler_buffer_size: 128,
+            per_connection_event_buffer_size: 8,
             dial_concurrency_factor: 8,
             max_connections_in: 16,
             max_connections_out: 16,
             max_connections_pending_in: 8,
             max_connections_pending_out: 8,
             max_connections_per_peer: 8,
-            mdns: true,
-            direct_node_addresses: Vec::new(),
-            notify_handler_buffer_size: 128,
-            per_connection_event_buffer_size: 8,
-            quic_port: 2022,
-            relay_mode: false,
-            relay_addresses: Vec::new(),
-            allow_peer_ids: AllowList::<PeerId>::Wildcard,
-            block_peer_ids: AllowList::<PeerId>::Wildcard,
         }
     }
 }